Updated solver unit tests
