+++ dummy.json
@@ -1,0 +1,1 @@
+{"model_type": "ib", "compound": "paracetamol", "dose_type": "c", "dose": 0.5, "dose_mass": null, "time_dose": null, "num_dose": null, "len_assay": 72.0, "len_interval": 0.25, "clearance": 0.4, "compartments": [[1.0, 0.3, "Main"], [0.5, 0.15, "Heart"]], "vis": "y", "curr_datetime": "1634833362.6048021"}
